+++ src/synthesizer/emissions/sed.py
@@ -13,6 +13,9 @@ Example usage:
     sed.get_photo_fnu(filters, nthreads=4)
 """
 
+import re
+
+import matplotlib.pyplot as plt
 import numpy as np
 from scipy.interpolate import interp1d
 from scipy.stats import linregress
@@ -33,19 +36,11 @@ from unyt import (
 
 from synthesizer import exceptions
 from synthesizer.conversions import lnu_to_llam
-from synthesizer.emissions import (
-    plot_observed_spectra,
-    plot_spectra,
-    plot_spectra_as_rainbow,
-)
 from synthesizer.extensions.timers import tic, toc
 from synthesizer.photometry import PhotometryCollection
 from synthesizer.synth_warnings import deprecated, warn
 from synthesizer.units import Quantity, accepts
-from synthesizer.utils import (
-    TableFormatter,
-    rebin_1d,
-)
+from synthesizer.utils import TableFormatter, rebin_1d, wavelength_to_rgba
 from synthesizer.utils.integrate import integrate_last_axis
 
 
@@ -1517,3 +1512,492 @@ class Sed:
         A wrapper for synthesizer.emissions.plot_spectra_as_rainbow()
         """
         return plot_spectra_as_rainbow(self, **kwargs)
+
+
+def plot_spectra(
+    spectra,
+    fig=None,
+    ax=None,
+    show=False,
+    ylimits=(),
+    xlimits=(),
+    figsize=(3.5, 5),
+    label=None,
+    draw_legend=True,
+    x_units=None,
+    y_units=None,
+    quantity_to_plot="lnu",
+):
+    """
+    Plot a spectra or dictionary of spectra.
+
+    Plots either a specific spectra or all spectra provided in a dictionary.
+    The plotted "type" of spectra is defined by the quantity_to_plot keyword
+    arrgument which defaults to "lnu".
+
+    This is a generic plotting function to be used either directly or to be
+    wrapped by helper methods through Synthesizer.
+
+    Args:
+        spectra (dict/Sed)
+            The Sed objects from which to plot. This can either be a dictionary
+            of Sed objects to plot multiple or a single Sed object to only plot
+            one.
+        fig (matplotlib.pyplot.figure)
+            The figure containing the axis. By default one is created in this
+            function.
+        ax (matplotlib.axes)
+            The axis to plot the data on. By default one is created in this
+            function.
+        show (bool)
+            Flag for whether to show the plot or just return the
+            figure and axes.
+        ylimits (tuple)
+            The limits to apply to the y axis. If not provided the limits
+            will be calculated with the lower limit set to 1000 (100) times
+            less than the peak of the spectrum for rest_frame (observed)
+            spectra.
+        xlimits (tuple)
+            The limits to apply to the x axis. If not provided the optimal
+            limits are found based on the ylimits.
+        figsize (tuple)
+            Tuple with size 2 defining the figure size.
+        label (string)
+            The label to give the spectra. Only applicable when Sed is a single
+            spectra.
+        draw_legend (bool)
+            Whether to draw the legend.
+        x_units (unyt.unit_object.Unit)
+            The units of the x axis. This will be converted to a string
+            and included in the axis label. By default the internal unit system
+            is assumed unless this is passed.
+        y_units (unyt.unit_object.Unit)
+            The units of the y axis. This will be converted to a string
+            and included in the axis label. By default the internal unit system
+            is assumed unless this is passed.
+        quantity_to_plot (string)
+            The sed property to plot. Can be "lnu", "luminosity" or "llam"
+            for rest frame spectra or "fnu", "flam" or "flux" for observed
+            spectra. Defaults to "lnu".
+
+    Returns:
+        fig (matplotlib.pyplot.figure)
+            The matplotlib figure object for the plot.
+        ax (matplotlib.axes)
+            The matplotlib axes object containing the plotted data.
+    """
+    # Check we have been given a valid quantity_to_plot
+    if quantity_to_plot not in (
+        "lnu",
+        "llam",
+        "luminosity",
+        "fnu",
+        "flam",
+        "flux",
+    ):
+        raise exceptions.InconsistentArguments(
+            f"{quantity_to_plot} is not a valid quantity_to_plot"
+            "(can be 'fnu' or 'flam')"
+        )
+
+    # Are we plotting in the rest_frame?
+    rest_frame = quantity_to_plot in ("lnu", "llam", "luminosity")
+
+    # Make a singular Sed a dictionary for ease below
+    if isinstance(spectra, Sed):
+        spectra = {
+            label if label is not None else "spectra": spectra,
+        }
+
+        # Don't draw a legend if not label given
+        if label is None and draw_legend:
+            warn("No label given, we will not draw a legend")
+            draw_legend = False
+
+    # If we don't already have a figure, make one
+    if fig is None:
+        # Set up the figure
+        fig = plt.figure(figsize=figsize)
+
+        # Define the axes geometry
+        left = 0.15
+        height = 0.6
+        bottom = 0.1
+        width = 0.8
+
+        # Create the axes
+        ax = fig.add_axes((left, bottom, width, height))
+
+        # Set the scale to log log
+        ax.loglog()
+
+    # Loop over the dict we have been handed, we want to do this backwards
+    # to ensure the most recent spectra are on top
+    keys = list(spectra.keys())[::-1]
+    seds = list(spectra.values())[::-1]
+    for key, sed in zip(keys, seds):
+        # Get the appropriate luminosity/flux and wavelengths
+        if rest_frame:
+            lam = sed.lam
+        else:
+            # Ensure we have fluxes
+            if sed.fnu is None:
+                raise exceptions.MissingSpectraType(
+                    f"This Sed has no fluxes ({key})! Have you called "
+                    "Sed.get_fnu()?"
+                )
+
+            # Ok everything is fine
+            lam = sed.obslam
+
+        plt_spectra = getattr(sed, quantity_to_plot)
+
+        # Prettify the label if not latex
+        if not any([c in key for c in ("$", "_")]):
+            key = key.replace("_", " ").title()
+
+        # Plot this spectra
+        ax.plot(lam, plt_spectra, lw=1, alpha=0.8, label=key)
+
+    # Do we not have y limtis?
+    if len(ylimits) == 0:
+        # Define initial xlimits
+        ylimits = [np.inf, -np.inf]
+
+        # Loop over spectra and get the total required limits
+        for sed in spectra.values():
+            # Get the maximum ignoring infinites
+            okinds = np.logical_and(
+                getattr(sed, quantity_to_plot) > 0,
+                getattr(sed, quantity_to_plot) < np.inf,
+            )
+            if True not in okinds:
+                continue
+            max_val = np.nanmax(getattr(sed, quantity_to_plot)[okinds])
+
+            # Derive the x limits
+            y_up = 10 ** (np.log10(max_val) * 1.05)
+            y_low = 10 ** (np.log10(max_val) - 5)
+
+            # Update limits
+            if y_low < ylimits[0]:
+                ylimits[0] = y_low
+            if y_up > ylimits[1]:
+                ylimits[1] = y_up
+
+    # Do we not have x limits?
+    if len(xlimits) == 0:
+        # Define initial xlimits
+        xlimits = [np.inf, -np.inf]
+
+        # Loop over spectra and get the total required limits
+        for sed in spectra.values():
+            # Derive the x limits from data above the ylimits
+            plt_spectra = getattr(sed, quantity_to_plot)
+            lam_mask = plt_spectra > ylimits[0]
+            if rest_frame:
+                lams_above = sed.lam[lam_mask]
+            else:
+                lams_above = sed.obslam[lam_mask]
+
+            # Saftey skip if no values are above the limit
+            if lams_above.size == 0:
+                continue
+
+            # Derive the x limits
+            x_low = 10 ** (np.log10(np.min(lams_above)) * 0.9)
+            x_up = 10 ** (np.log10(np.max(lams_above)) * 1.1)
+
+            # Update limits
+            if x_low < xlimits[0]:
+                xlimits[0] = x_low
+            if x_up > xlimits[1]:
+                xlimits[1] = x_up
+
+    # Set the limits
+    if not np.isnan(xlimits[0]) and not np.isnan(xlimits[1]):
+        ax.set_xlim(*xlimits)
+    if not np.isnan(ylimits[0]) and not np.isnan(ylimits[1]):
+        ax.set_ylim(*ylimits)
+
+    # Make the legend
+    if draw_legend and any(ax.get_legend_handles_labels()[1]):
+        ax.legend(fontsize=8, labelspacing=0.0)
+
+    # Parse the units for the labels and make them pretty
+    if x_units is None:
+        x_units = lam.units.latex_repr
+    else:
+        x_units = str(x_units)
+    if y_units is None:
+        y_units = plt_spectra.units.latex_repr
+    else:
+        y_units = str(y_units)
+
+    # Replace any \frac with a \ division
+    pattern = r"\{(.*?)\}\{(.*?)\}"
+    replacement = r"\1 \ / \ \2"
+    x_units = re.sub(pattern, replacement, x_units).replace(r"\frac", "")
+    y_units = re.sub(pattern, replacement, y_units).replace(r"\frac", "")
+
+    # Label the x axis
+    if rest_frame:
+        ax.set_xlabel(r"$\lambda/[\mathrm{" + x_units + r"}]$")
+    else:
+        ax.set_xlabel(r"$\lambda_\mathrm{obs}/[\mathrm{" + x_units + r"}]$")
+
+    # Label the y axis handling all possibilities
+    if quantity_to_plot == "lnu":
+        ax.set_ylabel(r"$L_{\nu}/[\mathrm{" + y_units + r"}]$")
+    elif quantity_to_plot == "llam":
+        ax.set_ylabel(r"$L_{\lambda}/[\mathrm{" + y_units + r"}]$")
+    elif quantity_to_plot == "luminosity":
+        ax.set_ylabel(r"$L/[\mathrm{" + y_units + r"}]$")
+    elif quantity_to_plot == "fnu":
+        ax.set_ylabel(r"$F_{\nu}/[\mathrm{" + y_units + r"}]$")
+    elif quantity_to_plot == "flam":
+        ax.set_ylabel(r"$F_{\lambda}/[\mathrm{" + y_units + r"}]$")
+    else:
+        ax.set_ylabel(r"$F/[\mathrm{" + y_units + r"}]$")
+
+    # Are we showing?
+    if show:
+        plt.show()
+
+    return fig, ax
+
+
+def plot_observed_spectra(
+    spectra,
+    redshift,
+    fig=None,
+    ax=None,
+    show=False,
+    ylimits=(),
+    xlimits=(),
+    figsize=(3.5, 5),
+    label=None,
+    draw_legend=True,
+    x_units=None,
+    y_units=None,
+    filters=None,
+    quantity_to_plot="fnu",
+):
+    """
+    Plot a spectra or dictionary of spectra.
+
+    Plots either a specific observed spectra or all observed spectra
+    provided in a dictionary.
+
+    This function is a wrapper around plot_spectra.
+
+    This is a generic plotting function to be used either directly or to be
+    wrapped by helper methods through Synthesizer.
+
+    Args:
+        spectra (dict/Sed)
+            The Sed objects from which to plot. This can either be a dictionary
+            of Sed objects to plot multiple or a single Sed object to only plot
+            one.
+        redshift (float)
+            The redshift of the observation.
+        fig (matplotlib.pyplot.figure)
+            The figure containing the axis. By default one is created in this
+            function.
+        ax (matplotlib.axes)
+            The axis to plot the data on. By default one is created in this
+            function.
+        show (bool)
+            Flag for whether to show the plot or just return the
+            figure and axes.
+        ylimits (tuple)
+            The limits to apply to the y axis. If not provided the limits
+            will be calculated with the lower limit set to 1000 (100) times
+            less than the peak of the spectrum for rest_frame (observed)
+            spectra.
+        xlimits (tuple)
+            The limits to apply to the x axis. If not provided the optimal
+            limits are found based on the ylimits.
+        figsize (tuple)
+            Tuple with size 2 defining the figure size.
+        label (string)
+            The label to give the spectra. Only applicable when Sed is a single
+            spectra.
+        draw_legend (bool)
+            Whether to draw the legend.
+        x_units (unyt.unit_object.Unit)
+            The units of the x axis. This will be converted to a string
+            and included in the axis label. By default the internal unit system
+            is assumed unless this is passed.
+        y_units (unyt.unit_object.Unit)
+            The units of the y axis. This will be converted to a string
+            and included in the axis label. By default the internal unit system
+            is assumed unless this is passed.
+        filters (FilterCollection)
+            If given then the photometry is computed and both the photometry
+            and filter curves are plotted
+        quantity_to_plot (string)
+            The sed property to plot. Can be "fnu", "flam", or "flux".
+            Defaults to "fnu".
+
+    Returns:
+        fig (matplotlib.pyplot.figure)
+            The matplotlib figure object for the plot.
+        ax (matplotlib.axes)
+            The matplotlib axes object containing the plotted data.
+    """
+    # Check we have been given a valid quantity_to_plot
+    if quantity_to_plot not in ("fnu", "flam"):
+        raise exceptions.InconsistentArguments(
+            f"{quantity_to_plot} is not a valid quantity_to_plot"
+            "(can be 'fnu' or 'flam')"
+        )
+
+    # Get the observed spectra plot
+    fig, ax = plot_spectra(
+        spectra,
+        fig=fig,
+        ax=ax,
+        show=False,
+        ylimits=ylimits,
+        xlimits=xlimits,
+        figsize=figsize,
+        label=label,
+        draw_legend=draw_legend,
+        x_units=x_units,
+        y_units=y_units,
+        quantity_to_plot=quantity_to_plot,
+    )
+
+    # Are we including photometry and filters?
+    if filters is not None:
+        # Add a filter axis
+        filter_ax = ax.twinx()
+        filter_ax.set_ylim(0, None)
+
+        # PLot each filter curve
+        for f in filters:
+            filter_ax.plot(f.lam * (1 + redshift), f.t)
+
+        # Make a singular Sed a dictionary for ease below
+        if isinstance(spectra, Sed):
+            spectra = {
+                label if label is not None else "spectra": spectra,
+            }
+
+        # Loop over spectra plotting photometry and filter curves
+        for sed in spectra.values():
+            # Get the photometry
+            sed.get_photo_fnu(filters)
+
+            # Plot the photometry for each filter
+            for f in filters:
+                piv_lam = f.pivwv()
+                ax.scatter(
+                    piv_lam * (1 + redshift),
+                    sed.photo_fnu[f.filter_code],
+                    zorder=4,
+                )
+
+    if show:
+        plt.show()
+
+    return fig, ax
+
+
+def plot_spectra_as_rainbow(
+    sed,
+    figsize=(5, 0.5),
+    lam_min=3000,
+    lam_max=8000,
+    include_xaxis=True,
+    logged=False,
+    min_log_lnu=-2.0,
+    use_fnu=False,
+):
+    """
+    Create a plot of the spectrum as a rainbow.
+
+    Arguments:
+        sed (synthesizer.emissions.Sed)
+            A synthesizer.emissions object.
+        figsize (tuple)
+            Fig-size tuple (width, height).
+        lam_min (float)
+            The min wavelength to plot in Angstroms.
+        lam_max (float)
+            The max wavelength to plot in Angstroms.
+        include_xaxis (bool)
+            Flag whther to include x-axis ticks and label.
+        logged (bool)
+            Flag whether to use logged luminosity.
+        min_log_lnu (float)
+            Minium luminosity to plot relative to the maximum.
+        use_fnu (bool)
+            Whether to plot fluxes or luminosities. If True
+            fluxes are plotted, otherwise luminosities.
+
+    Returns:
+        fig (matplotlib.pyplot.figure)
+            The matplotlib figure object for the plot.
+        ax (matplotlib.axes)
+            The matplotlib axes object containing the plotted data.
+    """
+    # Take sum of Seds if two dimensional
+    sed = sed.sum()
+
+    if use_fnu:
+        # Define filter for spectra
+        wavelength_indices = np.logical_and(
+            sed._obslam < lam_max, sed._obslam > lam_min
+        )
+        lam = sed.obslam[wavelength_indices].to("nm").value
+        spectra = sed._fnu[wavelength_indices]
+    else:
+        # define filter for spectra
+        wavelength_indices = np.logical_and(
+            sed._lam < lam_max, sed._lam > lam_min
+        )
+        lam = sed.lam[wavelength_indices].to("nm").value
+        spectra = sed._lnu[wavelength_indices]
+
+    # Normalise spectrum
+    spectra /= np.max(spectra)
+
+    # If logged rescale to between 0 and 1 using min_log_lnu
+    if logged:
+        spectra = (np.log10(spectra) - min_log_lnu) / (-min_log_lnu)
+        spectra[spectra < min_log_lnu] = 0
+
+    # initialise figure
+    fig = plt.figure(figsize=figsize)
+
+    # Initialise axes
+    if include_xaxis:
+        ax = fig.add_axes((0, 0.3, 1, 1))
+        ax.set_xlabel(r"$\lambda/\AA$")
+    else:
+        ax = fig.add_axes((0, 0.0, 1, 1))
+        ax.set_xticks([])
+
+    # Set background
+    ax.set_facecolor("black")
+
+    # Always turn off y-ticks
+    ax.set_yticks([])
+
+    # Get an array of colours
+    colours = np.array(
+        [
+            wavelength_to_rgba(lam_, alpha=spectra_)
+            for lam_, spectra_ in zip(lam, spectra)
+        ]
+    )
+
+    # Expand dimensions to get an image array
+    im = np.expand_dims(colours, axis=0)
+
+    # Show image
+    ax.imshow(im, aspect="auto", extent=(lam_min, lam_max, 0, 1))
+
+    return fig, ax
